Move fetching of packages to its own method
